requirements.txt
line-bot-sdk
flask
pythainlp
keras
pymysql
gunicorn==19.0.0
tensorflow
json